requirements.txt
boto3
pyautogui
Pillow
